Adding factory pattern to java
