--- Creational/Abstrac Factory/java/Vehicle.java
@@ -1,3 +1,0 @@
-public interface Vehicle {
-    void drive();
-}
