+++ Creational/AbstracFactory/java/Vehicle.java
@@ -1,0 +1,3 @@
+public interface Vehicle {
+    void drive();
+}
