+++ Creational/Factory/README.md
@@ -1,0 +1,7 @@
+# Factory 
+
+Factory pattern is one of the most used design patterns in Java. This type of design pattern comes under creational pattern as this pattern provides one of the best ways to create an object.
+
+In Factory pattern, we create object without exposing the creation logic to the client and refer to newly created object using a common interface.
+
+Text and Java code used can be found in [Tutorials Point WebSite](https://www.tutorialspoint.com/design_pattern/factory_pattern.htm)
