+++ Creational/Factory/java/Circle.java
@@ -1,0 +1,7 @@
+public class Circle implements Shape {
+
+    @Override
+    public void draw() {
+        System.out.println("Inside Circle::draw() method.");
+    }
+}
