+++ Creational/Factory/java/Rectangle.java
@@ -1,0 +1,7 @@
+public class Rectangle implements Shape {
+
+    @Override
+    public void draw() {
+        System.out.println("Inside Rectangle::draw() method.");
+    }
+}
