+++ Creational/Factory/java/Shape.java
@@ -1,0 +1,3 @@
+public interface Shape {
+    void draw();
+}
