+++ Creational/Factory/java/Square.java
@@ -1,0 +1,7 @@
+public class Square implements Shape {
+
+    @Override
+    public void draw() {
+        System.out.println("Inside Square::draw() method.");
+    }
+}
